\name{bkRNA18}
\alias{bkRNA18}
\docType{data}
\title{bkRNA18}
\description{
Please refer to the file /inst/doc/readme.pdf.
}
\usage{data("bkRNA18")}
\format{
  A data frame with 57955 observations on the following 18 variables.
  \describe{
    \item{\code{col36l6_1}}{a numeric vector}
    \item{\code{col38l6_3}}{a numeric vector}
    \item{\code{col39l6_5}}{a numeric vector}
    \item{\code{col40l6_7}}{a numeric vector}
    \item{\code{col44l6_9}}{a numeric vector}
    \item{\code{col45l6_11}}{a numeric vector}
    \item{\code{col47l6_13}}{a numeric vector}
    \item{\code{col48l6_97}}{a numeric vector}
    \item{\code{col52l6_17}}{a numeric vector}
    \item{\code{col36l6_2}}{a numeric vector}
    \item{\code{col38l6_4}}{a numeric vector}
    \item{\code{col39l6_6}}{a numeric vector}
    \item{\code{col40l6_8}}{a numeric vector}
    \item{\code{col44l6_10}}{a numeric vector}
    \item{\code{col45l6_12}}{a numeric vector}
    \item{\code{col47l6_14}}{a numeric vector}
    \item{\code{col48l6_98}}{a numeric vector}
    \item{\code{col52l6_18}}{a numeric vector}
  }
}
\examples{
data(bkRNA18)
## maybe str(bkRNA18) ; plot(bkRNA18) ...
}
\keyword{datasets}
