\name{bkRNA18_factors}
\alias{bkRNA18_factors}
\docType{data}
\title{bkRNA18_factors}
\description{
Please refer to the file /inst/doc/readme.pdf.
}
\usage{data("bkRNA18_factors")}
\format{
  A data frame with 18 observations on the following 13 variables.
  \describe{
    \item{\code{HG7}}{a numeric vector}
    \item{\code{ERCC}}{a numeric vector}
    \item{\code{TN}}{a numeric vector}
    \item{\code{TC}}{a numeric vector}
    \item{\code{CR}}{a numeric vector}
    \item{\code{NR}}{a numeric vector}
    \item{\code{DESeq}}{a numeric vector}
    \item{\code{UQ}}{a numeric vector}
    \item{\code{TMM}}{a numeric vector}
    \item{\code{TU}}{a numeric vector}
    \item{\code{NCS}}{a numeric vector}
    \item{\code{ES}}{a numeric vector}
    \item{\code{GAPDH}}{a numeric vector}
  }
}
\examples{
data(bkRNA18_factors)
## maybe str(bkRNA18_factors) ; plot(bkRNA18_factors) ...
}
\keyword{datasets}
