\name{scRNA663}
\alias{scRNA663}
\docType{data}
\title{scRNA663}
\description{
Please refer to the file /inst/doc/readme.pdf.
}
\usage{data("scRNA663")}
\format{
  A data frame with 57955 observations on the following 663 variables.
  \describe{
    \item{\code{col36l_1}}{a numeric vector}
    \item{\code{col36l_2}}{a numeric vector}
    \item{\code{col36l_3}}{a numeric vector}
    \item{\code{col36l_4}}{a numeric vector}
    \item{\code{col36l_5}}{a numeric vector}
    \item{\code{col36l_6}}{a numeric vector}
    \item{\code{col36l_7}}{a numeric vector}
    \item{\code{col36l_8}}{a numeric vector}
    \item{\code{col36l_9}}{a numeric vector}
    \item{\code{col36l_10}}{a numeric vector}
    \item{\code{col36l_11}}{a numeric vector}
    \item{\code{col36l_12}}{a numeric vector}
    \item{\code{col36l_13}}{a numeric vector}
    \item{\code{col36l_14}}{a numeric vector}
    \item{\code{col36l_15}}{a numeric vector}
    \item{\code{col36l_16}}{a numeric vector}
    \item{\code{col36l_17}}{a numeric vector}
    \item{\code{col36l_18}}{a numeric vector}
    \item{\code{col36l_19}}{a numeric vector}
    \item{\code{col36l_20}}{a numeric vector}
    \item{\code{col36l_21}}{a numeric vector}
    \item{\code{col36l_22}}{a numeric vector}
    \item{\code{col36l_23}}{a numeric vector}
    \item{\code{col36l_24}}{a numeric vector}
    \item{\code{col36l_25}}{a numeric vector}
    \item{\code{col36l_26}}{a numeric vector}
    \item{\code{col36l_27}}{a numeric vector}
    \item{\code{col36l_28}}{a numeric vector}
    \item{\code{col36l_29}}{a numeric vector}
    \item{\code{col36l_30}}{a numeric vector}
    \item{\code{col36l_31}}{a numeric vector}
    \item{\code{col36l_32}}{a numeric vector}
    \item{\code{col36l_33}}{a numeric vector}
    \item{\code{col36l_34}}{a numeric vector}
    \item{\code{col36l_35}}{a numeric vector}
    \item{\code{col36l_36}}{a numeric vector}
    \item{\code{col36l_37}}{a numeric vector}
    \item{\code{col36l_38}}{a numeric vector}
    \item{\code{col36l_39}}{a numeric vector}
    \item{\code{col36l_40}}{a numeric vector}
    \item{\code{col36l_41}}{a numeric vector}
    \item{\code{col36l_42}}{a numeric vector}
    \item{\code{col36l_43}}{a numeric vector}
    \item{\code{col36l_44}}{a numeric vector}
    \item{\code{col36l_45}}{a numeric vector}
    \item{\code{col36l_46}}{a numeric vector}
    \item{\code{col36l_47}}{a numeric vector}
    \item{\code{col36l_48}}{a numeric vector}
    \item{\code{col36l_49}}{a numeric vector}
    \item{\code{col36l_50}}{a numeric vector}
    \item{\code{col36l_51}}{a numeric vector}
    \item{\code{col36l_52}}{a numeric vector}
    \item{\code{col36l_53}}{a numeric vector}
    \item{\code{col36l_54}}{a numeric vector}
    \item{\code{col36l_55}}{a numeric vector}
    \item{\code{col36l_56}}{a numeric vector}
    \item{\code{col36l_57}}{a numeric vector}
    \item{\code{col36l_58}}{a numeric vector}
    \item{\code{col36l_59}}{a numeric vector}
    \item{\code{col36l_60}}{a numeric vector}
    \item{\code{col36l_61}}{a numeric vector}
    \item{\code{col36l_62}}{a numeric vector}
    \item{\code{col36l_63}}{a numeric vector}
    \item{\code{col36l_64}}{a numeric vector}
    \item{\code{col36l_65}}{a numeric vector}
    \item{\code{col36l_66}}{a numeric vector}
    \item{\code{col36l_67}}{a numeric vector}
    \item{\code{col36l_68}}{a numeric vector}
    \item{\code{col36l_69}}{a numeric vector}
    \item{\code{col36l_70}}{a numeric vector}
    \item{\code{col36l_71}}{a numeric vector}
    \item{\code{col38l_1}}{a numeric vector}
    \item{\code{col38l_2}}{a numeric vector}
    \item{\code{col38l_6}}{a numeric vector}
    \item{\code{col38l_7}}{a numeric vector}
    \item{\code{col38l_8}}{a numeric vector}
    \item{\code{col38l_10}}{a numeric vector}
    \item{\code{col38l_11}}{a numeric vector}
    \item{\code{col38l_12}}{a numeric vector}
    \item{\code{col38l_13}}{a numeric vector}
    \item{\code{col38l_14}}{a numeric vector}
    \item{\code{col38l_15}}{a numeric vector}
    \item{\code{col38l_16}}{a numeric vector}
    \item{\code{col38l_17}}{a numeric vector}
    \item{\code{col38l_19}}{a numeric vector}
    \item{\code{col38l_20}}{a numeric vector}
    \item{\code{col38l_21}}{a numeric vector}
    \item{\code{col38l_22}}{a numeric vector}
    \item{\code{col38l_23}}{a numeric vector}
    \item{\code{col38l_24}}{a numeric vector}
    \item{\code{col38l_25}}{a numeric vector}
    \item{\code{col38l_26}}{a numeric vector}
    \item{\code{col38l_27}}{a numeric vector}
    \item{\code{col38l_28}}{a numeric vector}
    \item{\code{col38l_29}}{a numeric vector}
    \item{\code{col38l_30}}{a numeric vector}
    \item{\code{col38l_31}}{a numeric vector}
    \item{\code{col38l_33}}{a numeric vector}
    \item{\code{col38l_34}}{a numeric vector}
    \item{\code{col38l_35}}{a numeric vector}
    \item{\code{col38l_36}}{a numeric vector}
    \item{\code{col38l_37}}{a numeric vector}
    \item{\code{col38l_39}}{a numeric vector}
    \item{\code{col38l_40}}{a numeric vector}
    \item{\code{col38l_41}}{a numeric vector}
    \item{\code{col38l_42}}{a numeric vector}
    \item{\code{col38l_43}}{a numeric vector}
    \item{\code{col38l_46}}{a numeric vector}
    \item{\code{col38l_47}}{a numeric vector}
    \item{\code{col38l_48}}{a numeric vector}
    \item{\code{col38l_49}}{a numeric vector}
    \item{\code{col38l_52}}{a numeric vector}
    \item{\code{col38l_55}}{a numeric vector}
    \item{\code{col38l_56}}{a numeric vector}
    \item{\code{col38l_57}}{a numeric vector}
    \item{\code{col38l_59}}{a numeric vector}
    \item{\code{col38l_60}}{a numeric vector}
    \item{\code{col38l_61}}{a numeric vector}
    \item{\code{col38l_62}}{a numeric vector}
    \item{\code{col38l_64}}{a numeric vector}
    \item{\code{col38l_65}}{a numeric vector}
    \item{\code{col38l_66}}{a numeric vector}
    \item{\code{col38l_67}}{a numeric vector}
    \item{\code{col38l_68}}{a numeric vector}
    \item{\code{col38l_69}}{a numeric vector}
    \item{\code{col38l_70}}{a numeric vector}
    \item{\code{col38l_72}}{a numeric vector}
    \item{\code{col39l1_47}}{a numeric vector}
    \item{\code{col39l1_48}}{a numeric vector}
    \item{\code{col39l1_49}}{a numeric vector}
    \item{\code{col39l1_50}}{a numeric vector}
    \item{\code{col39l1_51}}{a numeric vector}
    \item{\code{col39l1_52}}{a numeric vector}
    \item{\code{col39l1_53}}{a numeric vector}
    \item{\code{col39l1_54}}{a numeric vector}
    \item{\code{col39l1_55}}{a numeric vector}
    \item{\code{col39l1_56}}{a numeric vector}
    \item{\code{col39l1_57}}{a numeric vector}
    \item{\code{col39l1_58}}{a numeric vector}
    \item{\code{col39l1_59}}{a numeric vector}
    \item{\code{col39l1_60}}{a numeric vector}
    \item{\code{col39l1_61}}{a numeric vector}
    \item{\code{col39l1_62}}{a numeric vector}
    \item{\code{col39l1_63}}{a numeric vector}
    \item{\code{col39l1_64}}{a numeric vector}
    \item{\code{col39l1_69}}{a numeric vector}
    \item{\code{col39l1_70}}{a numeric vector}
    \item{\code{col39l1_71}}{a numeric vector}
    \item{\code{col39l1_72}}{a numeric vector}
    \item{\code{col39l1_73}}{a numeric vector}
    \item{\code{col39l1_74}}{a numeric vector}
    \item{\code{col39l1_75}}{a numeric vector}
    \item{\code{col39l1_76}}{a numeric vector}
    \item{\code{col39l1_77}}{a numeric vector}
    \item{\code{col39l1_78}}{a numeric vector}
    \item{\code{col39l1_79}}{a numeric vector}
    \item{\code{col39l1_80}}{a numeric vector}
    \item{\code{col39l1_81}}{a numeric vector}
    \item{\code{col39l1_82}}{a numeric vector}
    \item{\code{col39l1_83}}{a numeric vector}
    \item{\code{col39l1_84}}{a numeric vector}
    \item{\code{col39l1_85}}{a numeric vector}
    \item{\code{col39l1_86}}{a numeric vector}
    \item{\code{col39l1_87}}{a numeric vector}
    \item{\code{col39l1_88}}{a numeric vector}
    \item{\code{col39l1_89}}{a numeric vector}
    \item{\code{col39l1_90}}{a numeric vector}
    \item{\code{col39l1_91}}{a numeric vector}
    \item{\code{col39l1_92}}{a numeric vector}
    \item{\code{col39l1_93}}{a numeric vector}
    \item{\code{col39l1_94}}{a numeric vector}
    \item{\code{col39l1_95}}{a numeric vector}
    \item{\code{col39l1_96}}{a numeric vector}
    \item{\code{col39l2_20}}{a numeric vector}
    \item{\code{col39l2_21}}{a numeric vector}
    \item{\code{col39l2_22}}{a numeric vector}
    \item{\code{col39l2_23}}{a numeric vector}
    \item{\code{col39l2_24}}{a numeric vector}
    \item{\code{col39l2_25}}{a numeric vector}
    \item{\code{col39l2_26}}{a numeric vector}
    \item{\code{col39l2_27}}{a numeric vector}
    \item{\code{col39l2_28}}{a numeric vector}
    \item{\code{col39l2_29}}{a numeric vector}
    \item{\code{col39l2_30}}{a numeric vector}
    \item{\code{col39l2_31}}{a numeric vector}
    \item{\code{col39l2_32}}{a numeric vector}
    \item{\code{col39l2_33}}{a numeric vector}
    \item{\code{col39l2_34}}{a numeric vector}
    \item{\code{col39l2_35}}{a numeric vector}
    \item{\code{col39l2_36}}{a numeric vector}
    \item{\code{col39l2_37}}{a numeric vector}
    \item{\code{col39l2_38}}{a numeric vector}
    \item{\code{col39l2_39}}{a numeric vector}
    \item{\code{col39l2_40}}{a numeric vector}
    \item{\code{col39l2_41}}{a numeric vector}
    \item{\code{col39l2_42}}{a numeric vector}
    \item{\code{col39l2_43}}{a numeric vector}
    \item{\code{col39l2_44}}{a numeric vector}
    \item{\code{col39l2_45}}{a numeric vector}
    \item{\code{col39l2_46}}{a numeric vector}
    \item{\code{col39l2_47}}{a numeric vector}
    \item{\code{col39l2_48}}{a numeric vector}
    \item{\code{col39l2_49}}{a numeric vector}
    \item{\code{col39l2_50}}{a numeric vector}
    \item{\code{col39l2_52}}{a numeric vector}
    \item{\code{col39l2_53}}{a numeric vector}
    \item{\code{col39l2_54}}{a numeric vector}
    \item{\code{col39l2_55}}{a numeric vector}
    \item{\code{col39l2_56}}{a numeric vector}
    \item{\code{col39l2_57}}{a numeric vector}
    \item{\code{col39l2_58}}{a numeric vector}
    \item{\code{col39l2_59}}{a numeric vector}
    \item{\code{col39l2_60}}{a numeric vector}
    \item{\code{col39l2_61}}{a numeric vector}
    \item{\code{col39l2_62}}{a numeric vector}
    \item{\code{col39l2_63}}{a numeric vector}
    \item{\code{col39l2_64}}{a numeric vector}
    \item{\code{col39l2_65}}{a numeric vector}
    \item{\code{col39l2_66}}{a numeric vector}
    \item{\code{col39l2_67}}{a numeric vector}
    \item{\code{col39l2_68}}{a numeric vector}
    \item{\code{col39l2_69}}{a numeric vector}
    \item{\code{col39l2_70}}{a numeric vector}
    \item{\code{col39l2_72}}{a numeric vector}
    \item{\code{col39l2_73}}{a numeric vector}
    \item{\code{col39l2_74}}{a numeric vector}
    \item{\code{col39l3_31}}{a numeric vector}
    \item{\code{col39l3_32}}{a numeric vector}
    \item{\code{col39l3_33}}{a numeric vector}
    \item{\code{col39l3_34}}{a numeric vector}
    \item{\code{col39l3_35}}{a numeric vector}
    \item{\code{col39l3_36}}{a numeric vector}
    \item{\code{col39l3_38}}{a numeric vector}
    \item{\code{col39l3_39}}{a numeric vector}
    \item{\code{col39l3_40}}{a numeric vector}
    \item{\code{col39l3_41}}{a numeric vector}
    \item{\code{col39l3_42}}{a numeric vector}
    \item{\code{col39l3_43}}{a numeric vector}
    \item{\code{col39l3_44}}{a numeric vector}
    \item{\code{col39l3_45}}{a numeric vector}
    \item{\code{col39l3_46}}{a numeric vector}
    \item{\code{col39l3_47}}{a numeric vector}
    \item{\code{col39l3_48}}{a numeric vector}
    \item{\code{col39l3_49}}{a numeric vector}
    \item{\code{col39l3_50}}{a numeric vector}
    \item{\code{col39l3_51}}{a numeric vector}
    \item{\code{col39l3_52}}{a numeric vector}
    \item{\code{col39l3_53}}{a numeric vector}
    \item{\code{col39l3_54}}{a numeric vector}
    \item{\code{col39l3_55}}{a numeric vector}
    \item{\code{col39l3_56}}{a numeric vector}
    \item{\code{col39l3_57}}{a numeric vector}
    \item{\code{col39l3_58}}{a numeric vector}
    \item{\code{col39l3_59}}{a numeric vector}
    \item{\code{col39l3_60}}{a numeric vector}
    \item{\code{col39l3_61}}{a numeric vector}
    \item{\code{col39l3_62}}{a numeric vector}
    \item{\code{col39l3_63}}{a numeric vector}
    \item{\code{col39l3_64}}{a numeric vector}
    \item{\code{col39l3_65}}{a numeric vector}
    \item{\code{col39l3_66}}{a numeric vector}
    \item{\code{col39l3_67}}{a numeric vector}
    \item{\code{col39l3_68}}{a numeric vector}
    \item{\code{col39l3_69}}{a numeric vector}
    \item{\code{col39l3_70}}{a numeric vector}
    \item{\code{col39l3_71}}{a numeric vector}
    \item{\code{col39l3_72}}{a numeric vector}
    \item{\code{col39l3_73}}{a numeric vector}
    \item{\code{col39l3_74}}{a numeric vector}
    \item{\code{col39l3_75}}{a numeric vector}
    \item{\code{col39l3_76}}{a numeric vector}
    \item{\code{col39l3_77}}{a numeric vector}
    \item{\code{col39l3_78}}{a numeric vector}
    \item{\code{col39l3_79}}{a numeric vector}
    \item{\code{col39l3_80}}{a numeric vector}
    \item{\code{col39l3_81}}{a numeric vector}
    \item{\code{col39l3_82}}{a numeric vector}
    \item{\code{col39l3_83}}{a numeric vector}
    \item{\code{col39l3_85}}{a numeric vector}
    \item{\code{col40l_1}}{a numeric vector}
    \item{\code{col40l_2}}{a numeric vector}
    \item{\code{col40l_3}}{a numeric vector}
    \item{\code{col40l_4}}{a numeric vector}
    \item{\code{col40l_5}}{a numeric vector}
    \item{\code{col40l_6}}{a numeric vector}
    \item{\code{col40l_7}}{a numeric vector}
    \item{\code{col40l_8}}{a numeric vector}
    \item{\code{col40l_9}}{a numeric vector}
    \item{\code{col40l_10}}{a numeric vector}
    \item{\code{col40l_11}}{a numeric vector}
    \item{\code{col40l_12}}{a numeric vector}
    \item{\code{col40l_13}}{a numeric vector}
    \item{\code{col40l_14}}{a numeric vector}
    \item{\code{col40l_15}}{a numeric vector}
    \item{\code{col40l_16}}{a numeric vector}
    \item{\code{col40l_17}}{a numeric vector}
    \item{\code{col40l_18}}{a numeric vector}
    \item{\code{col40l_19}}{a numeric vector}
    \item{\code{col40l_20}}{a numeric vector}
    \item{\code{col40l_21}}{a numeric vector}
    \item{\code{col40l_22}}{a numeric vector}
    \item{\code{col40l_23}}{a numeric vector}
    \item{\code{col40l_24}}{a numeric vector}
    \item{\code{col40l_25}}{a numeric vector}
    \item{\code{col40l_26}}{a numeric vector}
    \item{\code{col40l_27}}{a numeric vector}
    \item{\code{col40l_28}}{a numeric vector}
    \item{\code{col40l_29}}{a numeric vector}
    \item{\code{col40l_30}}{a numeric vector}
    \item{\code{col40l_31}}{a numeric vector}
    \item{\code{col40l_32}}{a numeric vector}
    \item{\code{col40l_33}}{a numeric vector}
    \item{\code{col40l_34}}{a numeric vector}
    \item{\code{col40l_35}}{a numeric vector}
    \item{\code{col40l_37}}{a numeric vector}
    \item{\code{col40l_38}}{a numeric vector}
    \item{\code{col40l_39}}{a numeric vector}
    \item{\code{col40l_40}}{a numeric vector}
    \item{\code{col40l_41}}{a numeric vector}
    \item{\code{col40l_42}}{a numeric vector}
    \item{\code{col40l_44}}{a numeric vector}
    \item{\code{col40l_45}}{a numeric vector}
    \item{\code{col40l_46}}{a numeric vector}
    \item{\code{col40l_47}}{a numeric vector}
    \item{\code{col40l_48}}{a numeric vector}
    \item{\code{col40l_49}}{a numeric vector}
    \item{\code{col40l_50}}{a numeric vector}
    \item{\code{col44l1_1}}{a numeric vector}
    \item{\code{col44l1_2}}{a numeric vector}
    \item{\code{col44l1_3}}{a numeric vector}
    \item{\code{col44l1_4}}{a numeric vector}
    \item{\code{col44l1_8}}{a numeric vector}
    \item{\code{col44l1_11}}{a numeric vector}
    \item{\code{col44l1_12}}{a numeric vector}
    \item{\code{col44l1_13}}{a numeric vector}
    \item{\code{col44l1_14}}{a numeric vector}
    \item{\code{col44l1_15}}{a numeric vector}
    \item{\code{col44l1_16}}{a numeric vector}
    \item{\code{col44l1_17}}{a numeric vector}
    \item{\code{col44l1_18}}{a numeric vector}
    \item{\code{col44l1_19}}{a numeric vector}
    \item{\code{col44l1_20}}{a numeric vector}
    \item{\code{col44l1_24}}{a numeric vector}
    \item{\code{col44l1_28}}{a numeric vector}
    \item{\code{col44l1_29}}{a numeric vector}
    \item{\code{col44l1_30}}{a numeric vector}
    \item{\code{col44l1_32}}{a numeric vector}
    \item{\code{col44l1_33}}{a numeric vector}
    \item{\code{col44l1_36}}{a numeric vector}
    \item{\code{col44l1_38}}{a numeric vector}
    \item{\code{col44l1_40}}{a numeric vector}
    \item{\code{col44l1_41}}{a numeric vector}
    \item{\code{col44l1_42}}{a numeric vector}
    \item{\code{col44l1_43}}{a numeric vector}
    \item{\code{col44l1_47}}{a numeric vector}
    \item{\code{col44l1_48}}{a numeric vector}
    \item{\code{col44l1_50}}{a numeric vector}
    \item{\code{col44l1_53}}{a numeric vector}
    \item{\code{col44l1_58}}{a numeric vector}
    \item{\code{col44l1_59}}{a numeric vector}
    \item{\code{col44l1_60}}{a numeric vector}
    \item{\code{col44l1_64}}{a numeric vector}
    \item{\code{col44l1_66}}{a numeric vector}
    \item{\code{col44l1_67}}{a numeric vector}
    \item{\code{col44l1_68}}{a numeric vector}
    \item{\code{col44l1_69}}{a numeric vector}
    \item{\code{col44l1_70}}{a numeric vector}
    \item{\code{col44l1_71}}{a numeric vector}
    \item{\code{col44l1_73}}{a numeric vector}
    \item{\code{col44l1_74}}{a numeric vector}
    \item{\code{col44l1_75}}{a numeric vector}
    \item{\code{col44l1_76}}{a numeric vector}
    \item{\code{col44l1_77}}{a numeric vector}
    \item{\code{col44l1_79}}{a numeric vector}
    \item{\code{col44l1_80}}{a numeric vector}
    \item{\code{col44l1_82}}{a numeric vector}
    \item{\code{col44l1_83}}{a numeric vector}
    \item{\code{col44l1_84}}{a numeric vector}
    \item{\code{col44l1_85}}{a numeric vector}
    \item{\code{col44l1_86}}{a numeric vector}
    \item{\code{col44l1_87}}{a numeric vector}
    \item{\code{col44l1_89}}{a numeric vector}
    \item{\code{col44l1_90}}{a numeric vector}
    \item{\code{col44l1_91}}{a numeric vector}
    \item{\code{col44l1_92}}{a numeric vector}
    \item{\code{col44l1_93}}{a numeric vector}
    \item{\code{col44l1_94}}{a numeric vector}
    \item{\code{col44l2_1}}{a numeric vector}
    \item{\code{col44l2_2}}{a numeric vector}
    \item{\code{col44l2_3}}{a numeric vector}
    \item{\code{col44l2_4}}{a numeric vector}
    \item{\code{col44l2_5}}{a numeric vector}
    \item{\code{col44l2_6}}{a numeric vector}
    \item{\code{col44l2_9}}{a numeric vector}
    \item{\code{col44l2_10}}{a numeric vector}
    \item{\code{col44l2_12}}{a numeric vector}
    \item{\code{col44l2_14}}{a numeric vector}
    \item{\code{col44l2_17}}{a numeric vector}
    \item{\code{col44l2_18}}{a numeric vector}
    \item{\code{col44l2_19}}{a numeric vector}
    \item{\code{col44l2_20}}{a numeric vector}
    \item{\code{col44l2_23}}{a numeric vector}
    \item{\code{col44l2_26}}{a numeric vector}
    \item{\code{col44l2_27}}{a numeric vector}
    \item{\code{col44l2_28}}{a numeric vector}
    \item{\code{col44l2_30}}{a numeric vector}
    \item{\code{col44l2_31}}{a numeric vector}
    \item{\code{col44l2_32}}{a numeric vector}
    \item{\code{col44l2_35}}{a numeric vector}
    \item{\code{col44l2_36}}{a numeric vector}
    \item{\code{col44l2_37}}{a numeric vector}
    \item{\code{col44l2_38}}{a numeric vector}
    \item{\code{col44l2_40}}{a numeric vector}
    \item{\code{col44l7_67}}{a numeric vector}
    \item{\code{col44l7_68}}{a numeric vector}
    \item{\code{col44l7_69}}{a numeric vector}
    \item{\code{col44l7_70}}{a numeric vector}
    \item{\code{col44l7_71}}{a numeric vector}
    \item{\code{col44l7_75}}{a numeric vector}
    \item{\code{col44l7_76}}{a numeric vector}
    \item{\code{col44l7_77}}{a numeric vector}
    \item{\code{col44l7_78}}{a numeric vector}
    \item{\code{col44l7_79}}{a numeric vector}
    \item{\code{col44l7_80}}{a numeric vector}
    \item{\code{col44l7_82}}{a numeric vector}
    \item{\code{col44l7_83}}{a numeric vector}
    \item{\code{col44l7_84}}{a numeric vector}
    \item{\code{col44l7_85}}{a numeric vector}
    \item{\code{col44l7_86}}{a numeric vector}
    \item{\code{col44l7_87}}{a numeric vector}
    \item{\code{col44l7_88}}{a numeric vector}
    \item{\code{col44l7_89}}{a numeric vector}
    \item{\code{col44l7_90}}{a numeric vector}
    \item{\code{col44l7_91}}{a numeric vector}
    \item{\code{col44l7_92}}{a numeric vector}
    \item{\code{col44l7_93}}{a numeric vector}
    \item{\code{col44l7_94}}{a numeric vector}
    \item{\code{col44l7_95}}{a numeric vector}
    \item{\code{col44l7_96}}{a numeric vector}
    \item{\code{col44l8_17}}{a numeric vector}
    \item{\code{col44l8_19}}{a numeric vector}
    \item{\code{col44l8_21}}{a numeric vector}
    \item{\code{col44l8_22}}{a numeric vector}
    \item{\code{col44l8_23}}{a numeric vector}
    \item{\code{col44l8_24}}{a numeric vector}
    \item{\code{col44l8_25}}{a numeric vector}
    \item{\code{col44l8_26}}{a numeric vector}
    \item{\code{col44l8_28}}{a numeric vector}
    \item{\code{col44l8_29}}{a numeric vector}
    \item{\code{col44l8_31}}{a numeric vector}
    \item{\code{col44l8_32}}{a numeric vector}
    \item{\code{col44l8_33}}{a numeric vector}
    \item{\code{col44l8_34}}{a numeric vector}
    \item{\code{col44l8_43}}{a numeric vector}
    \item{\code{col44l8_45}}{a numeric vector}
    \item{\code{col44l8_46}}{a numeric vector}
    \item{\code{col44l8_47}}{a numeric vector}
    \item{\code{col44l8_48}}{a numeric vector}
    \item{\code{col44l8_50}}{a numeric vector}
    \item{\code{col44l8_51}}{a numeric vector}
    \item{\code{col44l8_52}}{a numeric vector}
    \item{\code{col44l8_53}}{a numeric vector}
    \item{\code{col44l8_54}}{a numeric vector}
    \item{\code{col44l8_55}}{a numeric vector}
    \item{\code{col44l8_56}}{a numeric vector}
    \item{\code{col44l8_58}}{a numeric vector}
    \item{\code{col44l8_59}}{a numeric vector}
    \item{\code{col44l8_60}}{a numeric vector}
    \item{\code{col44l8_61}}{a numeric vector}
    \item{\code{col44l8_62}}{a numeric vector}
    \item{\code{col44l8_63}}{a numeric vector}
    \item{\code{col44l8_67}}{a numeric vector}
    \item{\code{col44l8_68}}{a numeric vector}
    \item{\code{col44l8_70}}{a numeric vector}
    \item{\code{col44l8_71}}{a numeric vector}
    \item{\code{col44l8_72}}{a numeric vector}
    \item{\code{col44l8_73}}{a numeric vector}
    \item{\code{col44l8_74}}{a numeric vector}
    \item{\code{col44l8_76}}{a numeric vector}
    \item{\code{col44l8_78}}{a numeric vector}
    \item{\code{col44l8_81}}{a numeric vector}
    \item{\code{col44l8_85}}{a numeric vector}
    \item{\code{col44l8_86}}{a numeric vector}
    \item{\code{col44l8_88}}{a numeric vector}
    \item{\code{col44l8_94}}{a numeric vector}
    \item{\code{col45l2_42}}{a numeric vector}
    \item{\code{col45l2_43}}{a numeric vector}
    \item{\code{col45l2_45}}{a numeric vector}
    \item{\code{col45l2_47}}{a numeric vector}
    \item{\code{col45l2_48}}{a numeric vector}
    \item{\code{col45l2_50}}{a numeric vector}
    \item{\code{col45l2_51}}{a numeric vector}
    \item{\code{col45l2_52}}{a numeric vector}
    \item{\code{col45l2_54}}{a numeric vector}
    \item{\code{col45l2_55}}{a numeric vector}
    \item{\code{col45l2_56}}{a numeric vector}
    \item{\code{col45l2_57}}{a numeric vector}
    \item{\code{col45l2_58}}{a numeric vector}
    \item{\code{col45l2_59}}{a numeric vector}
    \item{\code{col45l2_60}}{a numeric vector}
    \item{\code{col45l2_61}}{a numeric vector}
    \item{\code{col45l2_62}}{a numeric vector}
    \item{\code{col45l2_63}}{a numeric vector}
    \item{\code{col45l2_64}}{a numeric vector}
    \item{\code{col45l2_65}}{a numeric vector}
    \item{\code{col45l2_66}}{a numeric vector}
    \item{\code{col45l2_67}}{a numeric vector}
    \item{\code{col45l2_68}}{a numeric vector}
    \item{\code{col45l2_69}}{a numeric vector}
    \item{\code{col45l2_70}}{a numeric vector}
    \item{\code{col45l2_71}}{a numeric vector}
    \item{\code{col45l2_76}}{a numeric vector}
    \item{\code{col45l2_77}}{a numeric vector}
    \item{\code{col45l2_78}}{a numeric vector}
    \item{\code{col45l2_81}}{a numeric vector}
    \item{\code{col45l2_83}}{a numeric vector}
    \item{\code{col45l2_84}}{a numeric vector}
    \item{\code{col45l2_85}}{a numeric vector}
    \item{\code{col45l2_88}}{a numeric vector}
    \item{\code{col45l2_89}}{a numeric vector}
    \item{\code{col45l2_90}}{a numeric vector}
    \item{\code{col45l2_91}}{a numeric vector}
    \item{\code{col45l2_92}}{a numeric vector}
    \item{\code{col45l2_93}}{a numeric vector}
    \item{\code{col45l2_94}}{a numeric vector}
    \item{\code{col45l2_95}}{a numeric vector}
    \item{\code{col45l7_24}}{a numeric vector}
    \item{\code{col45l7_25}}{a numeric vector}
    \item{\code{col45l8_2}}{a numeric vector}
    \item{\code{col45l8_3}}{a numeric vector}
    \item{\code{col45l8_4}}{a numeric vector}
    \item{\code{col45l8_5}}{a numeric vector}
    \item{\code{col45l8_6}}{a numeric vector}
    \item{\code{col45l8_7}}{a numeric vector}
    \item{\code{col45l8_8}}{a numeric vector}
    \item{\code{col45l8_9}}{a numeric vector}
    \item{\code{col45l8_11}}{a numeric vector}
    \item{\code{col45l8_12}}{a numeric vector}
    \item{\code{col45l8_13}}{a numeric vector}
    \item{\code{col45l8_14}}{a numeric vector}
    \item{\code{col45l8_35}}{a numeric vector}
    \item{\code{col47l7_26}}{a numeric vector}
    \item{\code{col47l7_27}}{a numeric vector}
    \item{\code{col47l7_28}}{a numeric vector}
    \item{\code{col47l7_29}}{a numeric vector}
    \item{\code{col47l7_30}}{a numeric vector}
    \item{\code{col47l7_31}}{a numeric vector}
    \item{\code{col47l7_32}}{a numeric vector}
    \item{\code{col47l7_33}}{a numeric vector}
    \item{\code{col47l7_34}}{a numeric vector}
    \item{\code{col47l7_35}}{a numeric vector}
    \item{\code{col47l7_36}}{a numeric vector}
    \item{\code{col47l7_37}}{a numeric vector}
    \item{\code{col47l7_38}}{a numeric vector}
    \item{\code{col47l7_41}}{a numeric vector}
    \item{\code{col47l7_42}}{a numeric vector}
    \item{\code{col47l7_44}}{a numeric vector}
    \item{\code{col47l7_45}}{a numeric vector}
    \item{\code{col47l7_47}}{a numeric vector}
    \item{\code{col47l7_48}}{a numeric vector}
    \item{\code{col47l7_49}}{a numeric vector}
    \item{\code{col47l7_50}}{a numeric vector}
    \item{\code{col47l7_51}}{a numeric vector}
    \item{\code{col47l7_54}}{a numeric vector}
    \item{\code{col47l7_57}}{a numeric vector}
    \item{\code{col47l7_58}}{a numeric vector}
    \item{\code{col47l7_59}}{a numeric vector}
    \item{\code{col47l7_60}}{a numeric vector}
    \item{\code{col47l7_63}}{a numeric vector}
    \item{\code{col47l7_64}}{a numeric vector}
    \item{\code{col47l7_65}}{a numeric vector}
    \item{\code{col47l7_66}}{a numeric vector}
    \item{\code{col48l6_2}}{a numeric vector}
    \item{\code{col48l6_4}}{a numeric vector}
    \item{\code{col48l6_5}}{a numeric vector}
    \item{\code{col48l6_6}}{a numeric vector}
    \item{\code{col48l6_7}}{a numeric vector}
    \item{\code{col48l6_8}}{a numeric vector}
    \item{\code{col48l6_9}}{a numeric vector}
    \item{\code{col48l6_10}}{a numeric vector}
    \item{\code{col48l6_11}}{a numeric vector}
    \item{\code{col48l6_12}}{a numeric vector}
    \item{\code{col48l6_13}}{a numeric vector}
    \item{\code{col48l6_14}}{a numeric vector}
    \item{\code{col48l6_16}}{a numeric vector}
    \item{\code{col48l6_17}}{a numeric vector}
    \item{\code{col48l6_18}}{a numeric vector}
    \item{\code{col48l6_19}}{a numeric vector}
    \item{\code{col48l6_20}}{a numeric vector}
    \item{\code{col48l6_21}}{a numeric vector}
    \item{\code{col48l6_22}}{a numeric vector}
    \item{\code{col48l6_24}}{a numeric vector}
    \item{\code{col48l6_26}}{a numeric vector}
    \item{\code{col48l6_27}}{a numeric vector}
    \item{\code{col48l6_28}}{a numeric vector}
    \item{\code{col48l6_30}}{a numeric vector}
    \item{\code{col48l6_31}}{a numeric vector}
    \item{\code{col48l6_34}}{a numeric vector}
    \item{\code{col48l6_35}}{a numeric vector}
    \item{\code{col48l6_36}}{a numeric vector}
    \item{\code{col48l6_39}}{a numeric vector}
    \item{\code{col48l6_40}}{a numeric vector}
    \item{\code{col48l6_42}}{a numeric vector}
    \item{\code{col48l6_43}}{a numeric vector}
    \item{\code{col48l6_44}}{a numeric vector}
    \item{\code{col48l6_46}}{a numeric vector}
    \item{\code{col48l6_47}}{a numeric vector}
    \item{\code{col48l6_48}}{a numeric vector}
    \item{\code{col48l6_50}}{a numeric vector}
    \item{\code{col48l6_51}}{a numeric vector}
    \item{\code{col48l6_52}}{a numeric vector}
    \item{\code{col48l6_53}}{a numeric vector}
    \item{\code{col48l6_54}}{a numeric vector}
    \item{\code{col48l6_55}}{a numeric vector}
    \item{\code{col48l6_56}}{a numeric vector}
    \item{\code{col48l6_58}}{a numeric vector}
    \item{\code{col48l6_60}}{a numeric vector}
    \item{\code{col48l6_61}}{a numeric vector}
    \item{\code{col48l6_62}}{a numeric vector}
    \item{\code{col48l6_63}}{a numeric vector}
    \item{\code{col48l6_64}}{a numeric vector}
    \item{\code{col48l6_65}}{a numeric vector}
    \item{\code{col48l6_66}}{a numeric vector}
    \item{\code{col48l6_67}}{a numeric vector}
    \item{\code{col48l6_68}}{a numeric vector}
    \item{\code{col48l6_69}}{a numeric vector}
    \item{\code{col48l6_70}}{a numeric vector}
    \item{\code{col48l6_71}}{a numeric vector}
    \item{\code{col48l6_72}}{a numeric vector}
    \item{\code{col48l6_73}}{a numeric vector}
    \item{\code{col48l6_74}}{a numeric vector}
    \item{\code{col48l6_75}}{a numeric vector}
    \item{\code{col48l6_76}}{a numeric vector}
    \item{\code{col48l6_77}}{a numeric vector}
    \item{\code{col48l6_78}}{a numeric vector}
    \item{\code{col48l6_79}}{a numeric vector}
    \item{\code{col48l6_80}}{a numeric vector}
    \item{\code{col48l6_81}}{a numeric vector}
    \item{\code{col48l6_82}}{a numeric vector}
    \item{\code{col48l6_85}}{a numeric vector}
    \item{\code{col48l6_86}}{a numeric vector}
    \item{\code{col48l6_88}}{a numeric vector}
    \item{\code{col48l6_90}}{a numeric vector}
    \item{\code{col48l6_92}}{a numeric vector}
    \item{\code{col48l6_93}}{a numeric vector}
    \item{\code{col48l6_94}}{a numeric vector}
    \item{\code{col48l6_95}}{a numeric vector}
    \item{\code{col48l6_96}}{a numeric vector}
    \item{\code{col48l7_1}}{a numeric vector}
    \item{\code{col48l7_2}}{a numeric vector}
    \item{\code{col48l7_3}}{a numeric vector}
    \item{\code{col48l7_4}}{a numeric vector}
    \item{\code{col48l7_5}}{a numeric vector}
    \item{\code{col48l7_7}}{a numeric vector}
    \item{\code{col48l7_9}}{a numeric vector}
    \item{\code{col48l7_11}}{a numeric vector}
    \item{\code{col48l7_12}}{a numeric vector}
    \item{\code{col48l7_13}}{a numeric vector}
    \item{\code{col48l7_15}}{a numeric vector}
    \item{\code{col48l7_16}}{a numeric vector}
    \item{\code{col48l7_18}}{a numeric vector}
    \item{\code{col48l7_20}}{a numeric vector}
    \item{\code{col48l7_21}}{a numeric vector}
  }
}
\examples{
data(scRNA663)
## maybe str(scRNA663) ; plot(scRNA663) ...
}
\keyword{datasets}
