\name{scRNA663_factors}
\alias{scRNA663_factors}
\docType{data}
\title{scRNA663_factors}
\description{
Please refer to the file /inst/doc/readme.pdf.
}
\usage{data("scRNA663_factors")}
\format{
  A data frame with 663 observations on the following 12 variables.
  \describe{
    \item{\code{HG7}}{a numeric vector}
    \item{\code{ERCC}}{a numeric vector}
    \item{\code{TN}}{a numeric vector}
    \item{\code{TC}}{a numeric vector}
    \item{\code{CR}}{a numeric vector}
    \item{\code{NR}}{a numeric vector}
    \item{\code{DESeq}}{a numeric vector}
    \item{\code{UQ}}{a numeric vector}
    \item{\code{TMM}}{a numeric vector}
    \item{\code{TU}}{a numeric vector}
    \item{\code{NCS}}{a numeric vector}
    \item{\code{ES}}{a numeric vector}
  }
}

\examples{
data(scRNA663_factors)
## maybe str(scRNA663_factors) ; plot(scRNA663_factors) ...
}
\keyword{datasets}
